drop the old (int) cast

+++ src/player/region.h
@@ -163,7 +163,7 @@ static void REGION_CopyRegion (wmfAPI* API,wmfRegion* dst,wmfRegion* src)
 
 		dst->type = src->type;
 
-		memcpy ((char*) dst->rects,(char*) src->rects,(int) (src->numRects * sizeof (wmfD_Rect)));
+		memcpy ((char*) dst->rects,(char*) src->rects,src->numRects * sizeof (wmfD_Rect));
 	}
 }
